% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdfmvna.r
\name{cdfmvna}
\alias{cdfmvna}
\title{The cumulative density function of the mvna}
\usage{
cdfmvna(a, r, s)
}
\arguments{
\item{a}{}

\item{r}{}

\item{s}{}
}
\description{
The cumulative density function of the mvna
}

