% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdfmvna.r
\name{pdfmvna}
\alias{pdfmvna}
\title{The probability density function of the mvna}
\usage{
pdfmvna(a, r, s)
}
\arguments{
\item{a}{}

\item{r}{}

\item{s}{}
}
\description{
The probability density function of the mvna
}

